requirements.txt
fastapi
uvicorn
debugpy
importlib-metadata
black  # for black
marshmallow
pymongo
python-dotenv
mysql-connector-python